Removed pytest fixtures to initialise sessions for each test
pytest fixtures to initialise sessions for each test to ensure db isolation and consistency after each test did not show behaviour as expected, possibly because calls to endpoints commit to the db and end current transaction preventing rollbacks
the fixture was replaced with fixtures that simply re-initialised the db and applied migrations on them in the current app_context passed to the test, but the code failed to give expected results for unknown reasons.
Currently, each test is written to ensure any commits to the db made in the test are rolled back before any assertions are done and the test is ended.
Possible future solutions include nested transaction (which would require the test body and endpoints to be re-written to use the nested_transaction fixture for commits) or creating a seperate test_db that is re-intialised after each test in the current app_context