Test added
Test added to verify handling of decorator when an invalid api endpoint is requested. A mock endpoint is created which is wrapped by the authenticate_principal decorator.
An invalid path is ppased to the request_context and the mock endpoint is called. mock_assert_found is used to capture calls to core.libs.assertions.assert_found and raise a FyleError.
We check whether mock_assert_found was called to verify if the exception was handled properly.